# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: KNFAPP API
  description: Backend API for VU Kaunas Faculty mobile app
  version: 1.7.0
  contact:
    name: VU KNF IT
servers:
  - url: /api
    description: API server
...
paths:
...
  /admin/invitations:
    post:
      tags: [admin]
      summary: Create invitation code
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [student, teacher, admin, curator]
                  default: student
                max_uses:
                  type: integer
                  default: 1
                expires_hours:
                  type: integer
                  default: 168
      responses:
        "201":
          description: Invitation created
    get:
      tags: [admin]
      summary: List invitation codes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Invitation list
  /admin/users:
    get:
      tags: [admin]
      summary: List all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list. Each user carries `active` (false = deactivated; such users cannot log in and their sessions are rejected)
  /admin/users/{userId}:
    patch:
      tags: [admin]
      summary: Update user role or active status
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [student, teacher, curator, admin]
                active:
                  type: boolean
      responses:
        "200":
          description: Updated user, including the resulting `active` flag
        "400":
          description: Invalid role or nothing to update
        "404":
          description: User not found
  /admin/stats:
    get:
      tags: [admin]
      summary: Admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats

  # ── Scraper ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...